+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `retain_back` (aka `truncate_front`) to `Deque`
 - Removed `impl Deref` for `HistoryBuf` to make accessing the raw backing array explicit (use `as_slice`).
 - Replace `HistoryBuf::as_slice` with `HistoryBuf::as_unordered_slice` to clarify ordering.
+- Replace `HistoryBuf::write` with `HistoryBuf::push` for a more consistent naming scheme.
 
 ## [v0.9.3] 2025-04-15